feat：弹幕场景礼物处理

+++ project/app/src/main/java/tech/sud/mgp/hello/ui/scenes/common/gift/model/GiftNotifyDetailModel.java
@@ -3,7 +3,7 @@ package tech.sud.mgp.hello.ui.scenes.common.gift.model;
 import tech.sud.mgp.hello.ui.scenes.base.model.UserInfo;
 
 public class GiftNotifyDetailModel {
-    public int giftID; // 礼物ID
+    public long giftID; // 礼物ID
     public int giftCount; // 礼物数量
     public GiftModel gift;
     public UserInfo sendUser;
